requirements.txt
requests==2.25.1
pyTelegramBotAPI==4.1.1
python-dotenv==0.19.1
